Moved sections into one scene and added index files for components
