+++ portfoliosite/src/App.js
@@ -1,20 +1,13 @@
 import React from 'react';
-import Header from './components/Header';
-import Hero from './components/Hero';
-import About from './components/About';
-import Projects from './components/Projects';
-import Contact from './components/Contact';
-import Footer from './components/Footer';
+import Home from './screens/Main/index';
+import { Header, Footer } from './components';
 import './App.css';
 
 function App() {
   return (
     <div>
       <Header />
-      <Hero />
-      <About />
-      <Projects />
-      <Contact />
+        <Home />
       <Footer />
     </div>
   );
