--- portfoliosite/src/components/About.css
@@ -1,18 +1,0 @@
-.about {
-  padding: 4rem 2rem;
-  text-align: center;
-  background: #1f1f1f; 
-}
-
-.about h2 {
-  font-size: 2.5rem;
-  margin-bottom: 1rem;
-  color: #ffffff; 
-}
-
-.about p {
-  font-size: 1.2rem;
-  max-width: 800px;
-  margin: 0 auto;
-  color: #e0e0e0; 
-}
