--- portfoliosite/src/components/About.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import './About.css';
-
-const About = () => {
-  return (
-    <section id="about" className="about">
-      <h2>About Me</h2>
-      <p>
-      Hey there! I’m a second-year student at the University of Huddersfield, where I’m constantly exploring how technology can solve problems and make life a little easier. Whether it’s coding, tinkering with new tools, or just figuring out how to automate the small stuff, I love finding creative ways to use tech to simplify everyday tasks. <br />
-      <br /> When I’m not buried in code or coursework, you’ll probably find me at the gym, reading a book or trying out some new tea.
-      </p>
-    </section>
-  );
-};
-
-export default About;
