--- portfoliosite/src/components/Contact.css
@@ -1,31 +1,0 @@
-.contact {
-  padding: 4rem 2rem;
-  text-align: center;
-  background: #1f1f1f; 
-}
-
-.contact h2 {
-  font-size: 2.5rem;
-  margin-bottom: 2rem;
-  color: #ffffff; 
-}
-
-.contact-links {
-  display: flex;
-  justify-content: center;
-  gap: 2rem;
-}
-
-.contact-links a {
-  color: #e0e0e0; 
-  text-decoration: none;
-  font-size: 1.2rem;
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  transition: color 0.3s;
-}
-
-.contact-links a:hover {
-  color: #646464; 
-}
