--- portfoliosite/src/components/Hero.css
@@ -1,50 +1,0 @@
-.hero {
-  height: 100vh;
-  background: url('https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fmedia.giphy.com%2Fmedia%2FWJSg7GXTu7bq%2Fgiphy.gif&f=1&nofb=1&ipt=e78f0b2c6e28bfb877af51d47c36ea61d3f3caa61a96829a8234ff8cb4c88890&ipo=images') repeat center center/cover;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  color: #fff;
-  position: relative;
-}
-
-.hero::before {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.5); 
-}
-
-.hero-content {
-  position: relative;
-  z-index: 1;
-}
-
-.hero-content h1 {
-  font-size: 4rem;
-  margin-bottom: 1rem;
-}
-
-.hero-content p {
-  font-size: 1.5rem;
-  margin-bottom: 2rem;
-}
-
-.hero-content button {
-  padding: 0.8rem 2rem;
-  font-size: 1rem;
-  background: #858585; 
-  color: #fff;
-  border: none;
-  cursor: pointer;
-  transition: background 0.3s;
-}
-
-.hero-content button:hover {
-  color:#9c9c9c;
-  background: #292d2e; 
-}
