--- portfoliosite/src/components/Hero.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import './Hero.css';
-
-const Hero = () => {
-  return (
-    <section id="hero" className="hero">
-      <div className="hero-content">
-        <h1>Hi, I'm Igor</h1>
-        <p>I am a Second year computer science student, <br/> Currently studying at the University of Huddersfield.</p>
-        <button>View My Work</button>
-      </div>
-    </section>
-  );
-};
-
-export default Hero;
