--- portfoliosite/src/components/Projects.css
@@ -1,45 +1,0 @@
-.projects {
-  padding: 4rem 2rem;
-  background: #1f1f1f; /* Dark gray background */
-  color: #e0e0e0; /* Light text */
-}
-
-.projects h2 {
-  text-align: center;
-  font-size: 2.5rem;
-  margin-bottom: 2rem;
-  color: #ffffff; /* White heading */
-}
-
-.project-list {
-  display: grid;
-  gap: 1.5rem;
-  max-width: 800px;
-  margin: 0 auto;
-}
-
-.project-card {
-  display: block;
-  background: #2c2c2c; /* Slightly lighter card background */
-  padding: 1.5rem;
-  border-radius: 8px;
-  text-decoration: none;
-  color: inherit;
-  transition: transform 0.3s, box-shadow 0.3s;
-}
-
-.project-card:hover {
-  transform: translateY(-3px);
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.3);
-}
-
-.project-card h3 {
-  color: #e0e0e0; 
-  margin-bottom: 0.5rem;
-  font-size: 1.5rem;
-}
-
-.project-card p {
-  color: #e0e0e0; 
-  line-height: 1.6;
-}
