+++ portfoliosite/src/components/index.js
@@ -1,0 +1,2 @@
+export { default as Header } from './Header/Header.js';
+export { default as Footer } from './Footer/Footer.js';
