+++ portfoliosite/src/screens/Main/index.css
@@ -1,0 +1,155 @@
+/* Hero Section */
+
+.hero {
+    height: 100vh;
+    background: url('https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fmedia.giphy.com%2Fmedia%2FWJSg7GXTu7bq%2Fgiphy.gif&f=1&nofb=1&ipt=e78f0b2c6e28bfb877af51d47c36ea61d3f3caa61a96829a8234ff8cb4c88890&ipo=images') repeat center center/cover;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    color: #fff;
+    position: relative;
+  }
+  
+  .hero::before {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.5); 
+  }
+  
+  .hero-content {
+    position: relative;
+    z-index: 1;
+  }
+  
+  .hero-content h1 {
+    font-size: 4rem;
+    margin-bottom: 1rem;
+  }
+  
+  .hero-content p {
+    font-size: 1.5rem;
+    margin-bottom: 2rem;
+  }
+  
+  .hero-content button {
+    padding: 0.8rem 2rem;
+    font-size: 1rem;
+    background: #858585; 
+    color: #fff;
+    border: none;
+    cursor: pointer;
+    transition: background 0.3s;
+  }
+  
+  .hero-content button:hover {
+    color:#9c9c9c;
+    background: #292d2e; 
+  }
+
+  /* About Section */
+
+  .about {
+    padding: 4rem 2rem;
+    text-align: center;
+    background: #1f1f1f; 
+  }
+  
+  .about h2 {
+    font-size: 2.5rem;
+    margin-bottom: 1rem;
+    color: #ffffff; 
+  }
+  
+  .about p {
+    font-size: 1.2rem;
+    max-width: 800px;
+    margin: 0 auto;
+    color: #e0e0e0; 
+  }
+
+  /* Project Section */
+
+  .projects {
+    padding: 4rem 2rem;
+    background: #1f1f1f;
+    color: #e0e0e0; 
+  }
+  
+  .projects h2 {
+    text-align: center;
+    font-size: 2.5rem;
+    margin-bottom: 2rem;
+    color: #ffffff; 
+  }
+  
+  .project-list {
+    display: grid;
+    gap: 1.5rem;
+    max-width: 800px;
+    margin: 0 auto;
+  }
+  
+  .project-card {
+    display: block;
+    background: #2c2c2c; 
+    padding: 1.5rem;
+    border-radius: 8px;
+    text-decoration: none;
+    color: inherit;
+    transition: transform 0.3s, box-shadow 0.3s;
+  }
+  
+  .project-card:hover {
+    transform: translateY(-3px);
+    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.3);
+  }
+  
+  .project-card h3 {
+    color: #e0e0e0; 
+    margin-bottom: 0.5rem;
+    font-size: 1.5rem;
+  }
+  
+  .project-card p {
+    color: #e0e0e0; 
+    line-height: 1.6;
+  }
+
+  /* Contact Section */
+
+  .contact {
+    padding: 4rem 2rem;
+    text-align: center;
+    background: #1f1f1f; 
+  }
+  
+  .contact h2 {
+    font-size: 2.5rem;
+    margin-bottom: 2rem;
+    color: #ffffff; 
+  }
+  
+  .contact-links {
+    display: flex;
+    justify-content: center;
+    gap: 2rem;
+  }
+  
+  .contact-links a {
+    color: #e0e0e0; 
+    text-decoration: none;
+    font-size: 1.2rem;
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+    transition: color 0.3s;
+  }
+  
+  .contact-links a:hover {
+    color: #646464; 
+  }
